Moved couple more ITs out to elastiknn-plugin/it
